+++ src/events/changeactivettexteditor.ts
@@ -1,3 +1,4 @@
+import { writeFile } from "fs-extra";
 import * as vscode from "vscode";
 import { settingsManager } from "../extension/settings.js";
 import { FileProperties } from "../miiservice/abstract/responsetypes.js";
@@ -8,6 +9,8 @@ import { CompareDocuments, OpenTextDocument, ShowConfirmPreviewMessage } from ".
 import { DownloadFile } from "../transfer/download.js";
 import { GetFileProperties } from "../transfer/request.js";
 import { filePropertiesTree } from "../ui/explorer/filepropertiestree.js";
+import logger from "../ui/logger.js";
+import statusBar, { Icon } from "../ui/statusbar.js";
 import { GetMainUserManager } from "../user/usermanager.js";
 import path = require("path");
 
@@ -43,12 +46,16 @@ async function CheckFileDifference(document: vscode.TextDocument, fileProp: File
         if (saidNoToTheseDocuments.find((no) => no.fsPath === document.uri.fsPath && no.modifiedTime == fileProp.Modified && no.modifiedUser == fileProp.ModifiedBy)) return;
 
         const file = await readFileService.call({ host: system.host, port: system.port }, sourcePath);
-        const payload = file.Rowsets.Rowset.Row.find((row) => row.Name == "Payload");
+        const payload = file?.Rowsets?.Rowset?.Row?.find((row) => row.Name == "Payload");
+        if (!payload) return;
+
         const remoteContent = Buffer.from(payload.Value, 'base64').toString();
         if (remoteContent != document.getText()) {
             const response = await ShowConfirmPreviewMessage("This file has been modified by " + modifiedUser + " at " + fileProp.Modified + ". Do you want to download it? File: " + fileProp.ObjectName);
             if (response === 1) {
-                DownloadFile(document.uri, userConfig, system);
+                await writeFile(document.fileName, Buffer.from(payload.Value, 'base64'), { encoding: "utf8" });
+                statusBar.updateBar("Downloaded " + path.basename(document.fileName), Icon.success, { duration: 3 });
+                logger.infos('Download File', path.basename(document.fileName) + ": Finished.");
             }
             else if (response === 2) {
                 const extension = path.extname(document.fileName).substring(1);