+++ src/miiservice/logoutservice.ts
@@ -1,21 +1,21 @@
 import { Request, Service } from './abstract/miiservice.js';
+const http = require('node:http');
+
 
 class LogOutService extends Service {
     name: string = "Log Out";
     mode: string = "XMII/Illuminator?service=logout";
+    //Works ME: manufacturing/logout
 
     async call({ host, port }: Request) {
-        // bug: This creates a session instead of destroying current
-        
-        /* 
         const url = this.get(host, port);
-        const { value, error, isError } = await this.fetch(new URL(url), false, null, 'none');
+        const { value, error, isError } = await this.fetch(new URL(url), { auth: false, convertResponse: 'none', redirect: 'manual' });
         if (!isError) {
             if (value.redirected && value.url == this.generateIP(host, port) + "/XMII/goService.jsp") {
                 return value;
             }
-            return null;  
-        }  */
+            return null;
+        }
     }
     get(host: string, port: number) {
         return this.generateURL(host, port, "http");