+++ CHANGELOG.md
@@ -4,15 +4,17 @@ All notable changes to the "miisync" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [0.10.1] - 12/06/2023
+## [0.11.0] - 04/07/2023
 
 ### Added
 
 - `Switch System` command. With this command you can switch the current main system.
+- `Upload Changes` command. With this command you can upload files that changed in git system.
 
 ### Fixed
 
 - Where 'Download/Upload Current File' command on explorer was disabled when there was no open editor.
+- Where logging out wouldn't destroy the current session.
 
 ## [0.10.0] - 12/06/2023